refactor: address further code review feedback
- Drop the PYGIT2_AVAILABLE branch from is_git_repository and
  something_to_commit (pygit2 is a hard dependency; the allow_unsafe
  subprocess path is kept because libgit2 ignores -c safe.directory).
- Remove performance-analysis asides and section-divider comments;
  de-capitalize "NOT" in the clone_bare_from_local docstring.
- Extract clear_default_branch() so resetting the attribute and the
  path-keyed cache always happen together.
- Log the public-key-cache patch miss at warning (it runs once, at
  import, so it cannot spam).
- Move the rsa_key_pair fixture into the test_handlers conftest.
